requirements.txt
fastapi==0.117.1
uvicorn==0.36.0
python-multipart==0.0.20
opencv-python==4.12.0.88
tensorflow==2.20.0
pillow==11.3.0
numpy==2.2.6